file_storage_servive/requirements.txt
flask
redis
flask-wtf
flask-sqlalchemy
